less code redundancy
