+++ tests/js/client/server_parameters/fill-block-cache-on-http-noncluster.js
@@ -26,7 +26,7 @@
 /// @author Copyright 2019, ArangoDB Inc, Cologne, Germany
 ////////////////////////////////////////////////////////////////////////////////
 
-const db = require('@arangodb').db;
+const helper = require('@arangodb/testutils/block-cache-test-helper');
 const cn = 'UnitTestsCollection';
 
 if (getOptions === true) {
@@ -37,16 +37,7 @@ if (getOptions === true) {
   };
 }
 if (runSetup === true) {
-  let payload = Array(1024).join("x"); 
-  let c = db._create(cn);
-  let docs = [];
-  for (let i = 0; i < 250 * 1000; ++i) {
-    docs.push({ _key: "test" + i, value1: i, value2: "test" + i, payload });
-    if (docs.length === 1000) {
-      c.insert(docs);
-      docs = [];
-    }
-  }
+  helper.setup(cn);
   return true;
 }
 
@@ -58,17 +49,7 @@ function FillBlockCacheSuite() {
   return {
     
     testHttpApi: function() {
-      const oldValue = db._engineStats()["rocksdb.block-cache-usage"];
-      arango.POST("/_api/cursor", { query: "FOR doc IN @@collection RETURN doc.value1", bindVars: { "@collection": cn }, options: { fillBlockCache: true } });
-      
-      const newValue = db._engineStats()["rocksdb.block-cache-usage"];
-      assertTrue(newValue >= oldValue + 100 * 1000 * 1000, { oldValue, newValue });
-      
-      arango.POST("/_api/cursor", { query: "FOR doc IN @@collection RETURN doc.value1", bindVars: { "@collection": cn }, options: { fillBlockCache: true } });
-      
-      // allow for some other background things to be run, thus add 1MB of leeway
-      const newValue2 = db._engineStats()["rocksdb.block-cache-usage"];
-      assertTrue(newValue2 <= newValue + 1000 * 1000, { newValue, newValue2 });
+      helper.testHttpApi(cn, true);
     },
 
   };
